Renamed the CLI module (in folder src/farn/cli) from 'farn.py' to '\_\_main\_\_.py'
